Fix computation of stock and simplify the SQL query.
The problem was that in the second subquery of the UNION the "from_location"
was used to join with "stock_location" table instead of using the "to_location"
field. That alone would have solved the issue but the query was still pretty
hard to read due to too many parenthesis and other details that could be
simplified.

Also, I moved the join with location from the subqueries to the main query
because:

- We prevent mistakes similar to the one solved by this patch
- I think it makes more sense to do the filtering in the outer layer because
  maybe in the future we decide that we want to see the stock of locations of
  type != storage. With the new implementation we only have change the query in
  one place.

By the way, the period cache subquery, was also applying the filter on a
different set of location types.